refactors server logic into server.js
