+++ .gitignore
@@ -73,7 +73,8 @@ web_modules/
 .env.test
 
 #firebase private key
-/firebase-service-account-file.json
+/firebaseServiceAccountKeyLive.json
+/firebaseServiceAccountKeyTesting.json
 
 # parcel-bundler cache (https://parceljs.org/)
 .cache
