+++ index.js
@@ -1,24 +1,11 @@
+//modules
 require("dotenv").config();
-const express = require("express");
-const path = require("path");
-const admin = require("firebase-admin");
-const serviceAccount = require(process.env.GOOGLE_APPLICATION_CREDENTIALS);
+const server = require("./server");
+
+//configs
 const PORT = process.env.PORT || 5000;
-const server = express();
-server.use(express.json());
-server.use(express.static(path.join(__dirname, "client/build")));
-admin.initializeApp({
-  credential: admin.credential.cert(serviceAccount),
-});
 
-server.get("/api", (req, res) => {
-  res.json({ message: "API is up!!!" });
-});
-
-server.get("*", (req, res) => {
-  res.sendFile(path.join(__dirname, "client/build", "index.html"));
-});
-
+//establish server connection on env port
 server.listen(PORT, () => {
   console.log(`\n*** Server listening on PORT:${PORT} ***\n`);
 });
