+++ package.json
@@ -18,7 +18,9 @@
   "dependencies": {
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
-    "firebase-admin": "^9.6.0"
+    "firebase-admin": "^9.6.0",
+    "helmet": "^4.6.0",
+    "knex": "^0.95.4"
   },
   "devDependencies": {
     "eslint": "^7.24.0",
